LPD-15847 portal-search-solr8-impl: Move fixture to the correct package
