+++ modules/apps/portal-search-solr8/portal-search-solr8-impl/src/test/java/com/liferay/portal/search/solr8/internal/legacy/query/StringQueryTest.java
@@ -5,7 +5,7 @@
 
 package com.liferay.portal.search.solr8.internal.legacy.query;
 
-import com.liferay.portal.search.solr8.internal.SolrIndexingFixture;
+import com.liferay.portal.search.solr8.internal.indexing.SolrIndexingFixture;
 import com.liferay.portal.search.test.util.indexing.IndexingFixture;
 import com.liferay.portal.search.test.util.legacy.query.BaseStringQueryTestCase;
 import com.liferay.portal.test.rule.LiferayUnitTestRule;
