+++ modules/apps/portal-search-solr8/portal-search-solr8-impl/src/test/java/com/liferay/portal/search/solr8/internal/logging/SolrIndexSearcherLoggingTest.java
@@ -8,8 +8,8 @@ package com.liferay.portal.search.solr8.internal.logging;
 import com.liferay.portal.kernel.search.generic.MatchAllQuery;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.search.solr8.internal.SolrIndexSearcher;
-import com.liferay.portal.search.solr8.internal.SolrIndexingFixture;
 import com.liferay.portal.search.solr8.internal.SolrUnitTestRequirements;
+import com.liferay.portal.search.solr8.internal.indexing.SolrIndexingFixture;
 import com.liferay.portal.search.solr8.internal.search.engine.adapter.search.CountSearchRequestExecutorImpl;
 import com.liferay.portal.search.solr8.internal.search.engine.adapter.search.SearchSearchRequestExecutorImpl;
 import com.liferay.portal.search.test.util.indexing.BaseIndexingTestCase;
