+++ modules/apps/portal-search-solr8/portal-search-solr8-impl/src/test/java/com/liferay/portal/search/solr8/internal/mappings/SolrDocumentMappingTest.java
@@ -5,7 +5,7 @@
 
 package com.liferay.portal.search.solr8.internal.mappings;
 
-import com.liferay.portal.search.solr8.internal.SolrIndexingFixture;
+import com.liferay.portal.search.solr8.internal.indexing.SolrIndexingFixture;
 import com.liferay.portal.search.test.util.indexing.IndexingFixture;
 import com.liferay.portal.search.test.util.mappings.BaseDocumentMappingTestCase;
 import com.liferay.portal.test.rule.LiferayUnitTestRule;
