+++ modules/apps/portal-search-solr8/portal-search-solr8-impl/src/test/java/com/liferay/portal/search/solr8/internal/query/string/QueryStringTest.java
@@ -5,7 +5,7 @@
 
 package com.liferay.portal.search.solr8.internal.query.string;
 
-import com.liferay.portal.search.solr8.internal.SolrIndexingFixture;
+import com.liferay.portal.search.solr8.internal.indexing.SolrIndexingFixture;
 import com.liferay.portal.search.test.util.indexing.IndexingFixture;
 import com.liferay.portal.search.test.util.query.string.BaseQueryStringTestCase;
 import com.liferay.portal.test.rule.LiferayUnitTestRule;
